Atualização!

+++ src/main/java/br/gov/cultura/DitelAdm/controller/UsuarioController.java
@@ -54,7 +54,7 @@ public class UsuarioController extends UrlController {
 	}
 
 	@RequestMapping(value = "{id}", method = RequestMethod.DELETE)
-	public String excluir(@PathVariable Long id, RedirectAttributes attributes) {
+	public String excluir(@PathVariable Integer id, RedirectAttributes attributes) {
 		cadastroUsuarioService.excluir(id);
 		attributes.addFlashAttribute("mensagem", "Cadastrado removido com sucesso!");
 		return "redirect:/inicio";
