+++ src/main/java/br/gov/cultura/DitelAdm/service/CadastroUsuarioService.java
@@ -5,10 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import br.gov.cultura.DitelAdm.model.Linha;
 import br.gov.cultura.DitelAdm.model.Usuario;
 import br.gov.cultura.DitelAdm.repository.Usuarios;
-import br.gov.cultura.DitelAdm.repository.filtro.CadastroFiltroPesquisa;
 
 @Service
 public class CadastroUsuarioService {
@@ -20,7 +18,7 @@ public class CadastroUsuarioService {
 		usuarios.save(usuario);
 	}
 
-	public void excluir(Long id) {
+	public void excluir(Integer id) {
 		usuarios.delete(id);
 	}
